% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{average_rlist}
\alias{average_rlist}
\title{Create an average raster from a raster list}
\usage{
average_rlist(flist, percentcov)
}
\arguments{
\item{flist}{character file.list}

\item{percentcov}{numeric tolerance to include a stacked pixel based on percent NA}
}
\description{
Create an average raster from a raster list
}
