% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{avmap}
\alias{avmap}
\title{create a difference map compared to average}
\usage{
avmap(yearmon, params, diffpath = NULL, avpath = NULL, percentcov = 0.6,
  tolerance = 1, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{survey of interest to compare against average}

\item{params}{variable name generally choice of "sal" or "chlext"}

\item{diffpath}{file.path to write output difference surface}

\item{avpath}{file.path to write output average surface}

\item{percentcov}{numeric account for the different raster extents by setting the percent of all surveys required before a pixel is included in difference from average computations}

\item{tolerance}{numeric number of months on either side of yearmon to include in the set of surfaces averaged . defaults to 1.}

\item{fdir}{character file path to local data directory}
}
\description{
takes a survey date as input and searches the DF_Surfaces folder for maps of the same parameter within the range of 1-2 months from yearmon for each year. These surfaces are averaged and compared to the surface from yearmon.
}
\examples{
\dontrun{
avmap(yearmon = 201505, params = "sal", diffpath = file.path(fdir,
"DF_Surfaces", yearmon, paste0("diff", params, ".tif")),
 percentcov = 0.6, tolerance = 1, fdir = fdir)
avmap(yearmon = 201505, params = "sal")
}
}

