% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chlmap.R
\name{chlmap}
\alias{chlmap}
\title{mapping chl data using existing surfaces and chl coeff}
\usage{
chlmap(yearmon, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric date in yyyymm format}
}
\description{
mapping chl data using existing surfaces and chl coeff
}
\examples{
chlmap(yearmon=201502)
}
\author{
Joseph Stachelek
}

