% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cond2sal}
\alias{cond2sal}
\title{Conductivity to salinity conversion}
\usage{
cond2sal(c, t = 25, P = 0)
}
\arguments{
\item{c}{numeric conducitivity in uS (measurements in mS need to multiplied by 1000)}

\item{t}{numeric celcius temperature}

\item{P}{numeric optional pressure (defaults to 0)}
}
\description{
Uses the PSS-78 practical salinity equation
}
\details{
algorithm based off an excel implementation by N. Iricanin
}
\examples{
\dontrun{csal<-cond2sal(c=dt$cond*1000,t=dt$temp)
plot(csal, dt$sal)
abline(a=0,b=1,col="red",lwd=2)}
}
\seealso{
\code{\link[wql]{ec2pss}}
}
