% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{create_rlist}
\alias{create_rlist}
\title{Create file listing of rasters from a date range and parameter name}
\usage{
create_rlist(rnge, params)
}
\arguments{
\item{rnge}{numeric string of 1, 2, or more dates dates in yyyymm format. A length 1 rnge will produce a single plot, a length 2 rnge will produce a series of plots bookended by the two dates, a rnge object with more than 2 dates will produce a series of plots exactly corresponding to the dates provided.}

\item{params}{character vector of parameter fields to plot legends and color ramps are defined for sal, chlext, and diffsal}
}
\description{
Create file listing of rasters from a date range and parameter name
}
\examples{
\dontrun{
create_rlist(rnge = 201509, params = "sal")
create_rlist(rnge = c(201509, 201605), params = c("sal", "salinity.pss", "salpsu"))
create_rlist(rnge = c(200808, 200910, 201002, 201004, 201007, 201102, 201105,
201206, 201209, 201212, 201305, 201308, 201311, 201404, 201407, 201410,
 201502, 201505, 201507, 201509), params = 'chlext')
 
}
}
