% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{cv_rlist}
\alias{cv_rlist}
\title{Create a variability raster from a raster list}
\usage{
cv_rlist(flist, percentcov)
}
\arguments{
\item{flist}{character file.list}

\item{percentcov}{numeric tolerance to include a stacked pixel based on percent NA}
}
\description{
Create a variability raster from a raster list. Divide sd by mean.
}
\examples{
\dontrun{
rlist <- create_rlist(good_years$x, "chlext")
cv_rlist(rlist$rlist, 0.8)
}
}
