% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grabclean.R
\name{grabclean}
\alias{grabclean}
\title{cleaning grab data}
\usage{
grabclean(yearmon, tofile = FALSE)
}
\arguments{
\item{yearmon}{numeric survey date in yyyymm format}

\item{tofile}{logical save output to file}
}
\description{
if column names do not match up make sure there is only one "date" column
}
\examples{
res<-grabclean(yearmon=201402,tofile=TRUE)
}

