% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grabget.R
\name{grabget}
\alias{grabget}
\title{aggregate grab data}
\usage{
grabget(rnge, remove.flags = FALSE, fdir = getOption("fdir"))
}
\arguments{
\item{rnge}{list of length two specifying date range in yyyymm format}

\item{remove.flags}{logical trim dataset based on flags?}

\item{fdir}{character file path to local data directory}
}
\description{
aggregate grab data
}
\examples{
\dontrun{
grabs<-grabget(rnge=c(201402,201410))
}
rnge<-201308
}

