% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grabplot.R
\name{grabplot}
\alias{grabplot}
\title{plotting grab data}
\usage{
grabplot(rnge = c(201407, 201410), params = c("sal", "chla"),
  plottype = "permitbarplot")
}
\description{
plotting grab data
}
\examples{
grabplot(rnge=201410,params=c("sal","chla"),plottype="permitbarplot")
}

