% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{logramp}
\alias{logramp}
\title{Create a log scaled color ramp for the grassmap function}
\usage{
logramp(n, maxrange)
}
\arguments{
\item{n}{integer number of breaks}

\item{maxrange}{integer maximum range
n <- 9
maxrange <- 20
logramp(n, maxrange)}
}
\description{
Create a log scaled color ramp for the grassmap function
}

