% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_dataflow.R
\name{solve_dataflow}
\alias{solve_dataflow}
\title{Solve the order of function calls in a flow of data}
\usage{
solve_dataflow(fun_name, envir)
}
\arguments{
\item{fun}{function at the end of the data flow}

\item{env}{environment of functions involved}
}
\value{
environment with function args replaced with call objects
}
\description{
Determines the order in which function dependencies need to be called such
that \code{fun}'s arguments are all resolved before it is called. A notion of a
"planning" environment is used (and ulimately returned), wherein function
calls from \code{env} are translated into synonymous call objects with their
arguments replaced by other function calls or symbols where appropriate.
}
