% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamget}
\alias{streamget}
\title{Retrieve previously cleaned full streaming datasets}
\usage{
streamget(yearmon)
}
\description{
Retrieve previously cleaned full streaming datasets
}
\examples{
yearmon<-201308
dt<-streamget(yearmon)
}

