% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streaminginterp.R
\name{streaminterp}
\alias{streaminterp}
\title{interpolation of streaming data}
\usage{
streaminterp(dt, paramlist, yearmon)
}
\arguments{
\item{dt}{input data frame}

\item{paramlist}{list of parameters (dt column names) to interpolate}

\item{yearmon}{a file path used to extract basename}
}
\description{
interpolation of streaming data
}
\examples{
dt<-streamget(yearmon=201502)
streaminterp(dt,paramlist=c("sal"),yearmon=201502)
}

