% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamparse}
\alias{streamparse}
\title{Parse old cleaned streaming files}
\usage{
streamparse(yearmon, tofile = FALSE, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric yyyymm date}

\item{tofile}{logical save to file?}

\item{fdir}{character file path to local data directory}
}
\description{
Parse old cleaned streaming files
}
\examples{
\dontrun{dt<-streamparse(yearmon=201002)}
}

