% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfget.R
\name{surfget}
\alias{surfget}
\title{Retrieve an interpolated Dataflow surface}
\usage{
surfget(rnge, param, fdir = getOption("fdir"))
}
\arguments{
\item{rnge}{list of length one or two specifying date range in yyyymm format}

\item{param}{character parameter name}

\item{fdir}{character file path to local data directory}
}
\description{
Retrieve an interpolated Dataflow surface from data directory
}
\examples{
\dontrun{ 
surfs <- surfget(rnge = c(201402, 201410), param = "chlext")
surfs <- surfget(rnge = 201509, param = "sal")
}
}
