% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validplatform.R
\name{validdbhydro}
\alias{validdbhydro}
\title{Validate Dataflow output against DBHYDRO data}
\usage{
validdbhydro(yearmon, params = "chlext", tolerance = 1,
  dbfile = file.path(), fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric 6 character date in YYYYMM format}

\item{params}{character string of parameter names}

\item{tolerance}{numeric number of days pre/post streaming to search for in DBHYDRO}

\item{dbfile}{file.path pointing at a cleaned database of dbhydro measurements}

\item{fdir}{file.path to data folder}
}
\description{
Validate Dataflow output against DBHYDRO data
}
\examples{
\dontrun{
dbfile <- file.path("/home","jose","Documents","Science","Data","ENP_MMN","WQSuM_2015.csv")
validdbhydro("201502",params="chlext",tolerance=10, dbfile=dbfile)
validdbhydro(201410,params="chlext",tolerance=13,dbfile=dbfile)
validdbhydro(201407,params="chlext",tolerance=2,dbfile=dbfile)
validdbhydro(201404,params="chlext",tolerance=17,dbfile=dbfile)
validdbhydro(201311,params="chlext",tolerance=10,dbfile=dbfile)
validdbhydro(201308,params="chlext",tolerance=5,dbfile=dbfile)
}
}

