% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validplatform.R
\name{validplatform}
\alias{validdbhydro}
\alias{validplatform}
\title{validate Dataflow output against platform data}
\usage{
validdbhydro(yearmon, params = "chlext", tolerance = 1,
  dbfile = file.path(), fdir = getOption("fdir"))
}
\arguments{
\item{tolerance}{numeric number of days pre/post streaming to search for in DBHYDRO}

\item{dbfile}{file.path pointing at a cleaned database of dbhydro measurements}
}
\description{
validate Dataflow output against platform data
}
\examples{
dbfile<-file.path("/home","jose","Documents","Science","Data","ENP_MMN","WQSuM_2015.csv")
validdbhydro("201502",params="chlext",tolerance=10, dbfile=dbfile)
validdbhydro(201410,params="chlext",tolerance=13,dbfile=dbfile)
validdbhydro(201407,params="chlext",tolerance=2,dbfile=dbfile)
validdbhydro(201404,params="chlext",tolerance=17,dbfile=dbfile)
validdbhydro(201311,params="chlext",tolerance=10,dbfile=dbfile)
validdbhydro(201308,params="chlext",tolerance=5,dbfile=dbfile)
}

