% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chlcoeff.R
\name{chlcoef}
\alias{chlcoef}
\title{calculation of extracted versus fluoresced chlorophyll coefficients}
\usage{
chlcoef(yearmon, remove.flags = FALSE)
}
\arguments{
\item{yearmon}{numeric date of survey}

\item{remove.flags}{logical trim dataset based on flags?}
}
\description{
calculation of extracted versus fluoresced chlorophyll coefficients
}
\details{
this function should be interactive
}

