% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chlcoeff.R
\name{chlcoef}
\alias{chlcoef}
\title{Calculation of extracted versus fluoresced chlorophyll coefficients}
\usage{
chlcoef(yearmon, varlist = NA, remove.flags = TRUE, overwrite = FALSE,
  fdir = getOption("fdir"), polypcut = 0.6, corcut = 0.7,
  streamcov = 0.5, checkvif = TRUE, logtransform = FALSE)
}
\arguments{
\item{yearmon}{numeric date of survey}

\item{varlist}{character vector of variable names to use in model}

\item{remove.flags}{logical trim dataset based on flags?}

\item{overwrite}{logical overwrite previous coefficients?}

\item{fdir}{file.path to data folder}

\item{polypcut}{numeric used to switch to a polynomial regression fit}

\item{corcut}{numeric used to screen bad variables with bad correlation coefficients}

\item{streamcov}{numeric percentage of non-missing streaming data used to exclude variables under consideration}

\item{checkvif}{logical check final equation for multicollinearity using VIF?}

\item{logtransform}{logical}
}
\description{
Calculate extracted versus fluoresced chlorophyll coefficients
}
\details{
this function should be interactive
}
\examples{
\dontrun{
chlcoef(201308)
chlcoef(201606, varlist = c("chl.ug.l", "turb.ntu", "phycoc", "c6cdom", "c6chl", "phycoe"), corcut = 0.75)
}
}
