% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chlmap.R
\name{chlmap}
\alias{chlmap}
\title{Create a chlorophyll concentration surface using streaming data and regression against extracted chlorophyll}
\usage{
chlmap(yearmon, remove.flags = TRUE, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric date in yyyymm format}

\item{remove.flags}{logical}

\item{fdir}{file.path to data folder}
}
\value{
An extracted chlorophyll surface and an updated FullDataSet file
}
\description{
Create a chlorophyll concentration surface using streaming data and regression against extracted chlorophyll
}
\details{
A new interpolation is run after calculating an extracted chlorophyll for all streaming observations. Calculated values that exceed the maximum observed grab sample concentration are discarded.
}
\examples{
\dontrun{
res<-chlmap(yearmon=201308)
}
}
\author{
Joseph Stachelek
}

