% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{coordinatize}
\alias{coordinatize}
\title{Convert georeferenced data.frames into projected SpatialPointsDataFrames}
\usage{
coordinatize(dt, latname = "latdec", lonname = "londec")
}
\arguments{
\item{dt}{data.frame with two coordinate columns}

\item{latname}{character column name of the "y" coordinate}

\item{lonname}{character column name of the "x" coordinate}
}
\description{
Convert georeferenced data.frames into projected SpatialPointsDataFrames
}
\examples{
\dontrun{
dt<-coordinatize(streamget(201002), latname = "lat_dd", lonname = "lon_dd")
}
}
