% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabget.R
\name{grabget}
\alias{grabget}
\title{Retrieve grab data}
\usage{
grabget(rnge, remove.flags = FALSE, fdir = getOption("fdir"))
}
\arguments{
\item{rnge}{list of length two specifying date range in yyyymm format}

\item{remove.flags}{logical trim dataset based on flags?}

\item{fdir}{character file path to local data directory}
}
\description{
Retrieve discrete grab sample data from data directory
}
\details{
Returns a list of data.frames (and a warning) if column names are not consistent among the surveys specified by rnge.
}
\examples{
\dontrun{ 
grabs<-grabget(rnge=c(201402,201410))
}
}
