% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabplot.R
\name{grabplot}
\alias{grabplot}
\title{Plotting grab data}
\usage{
grabplot(rnge = c(201407, 201410), params = c("sal", "chla"),
  plottype = "permitbarplot")
}
\arguments{
\item{rnge}{list of length two specifying date range in yyyymm format}

\item{params}{character list of parameters to plot}

\item{plottype}{character choice of plot type}
}
\description{
Plot discrete grab data
}
\examples{
\dontrun{
grabplot(rnge=201410,params=c("sal","chla"),plottype="permitbarplot")
}
}

