% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{grassmap}
\alias{grassmap}
\title{Publication quality maps with GRASS}
\usage{
grassmap(rnge = c(201502), params = c("sal"), fdir = getOption("fdir"),
  basin = "full", panel = FALSE, cleanup = TRUE, rotated = TRUE)
}
\arguments{
\item{rnge}{numeric string of no more than two dates in yyyymm format}

\item{params}{character. string of parameter fields to plot}

\item{fdir}{character file path to local data directory}

\item{panel}{logical. Combine output into a figure panel?}

\item{cleanup}{logical remove intermediate rasters and shapefiles?}

\item{rotated}{logical rotate canvas to fit Florida Bay more squarely? This requires the i.rotate extension to be installed and addons configured.}
}
\value{
output plots to the QGIS_plotting folder
}
\description{
not finished yet
}
\details{
probably need to implement this as a seperate package
}
\examples{
\dontrun{
grassmap(rnge=c(201505),params=c("sal"),basin="Manatee Bay")
grassmap(rnge=c(200707),params=c("sal"))

#create a new color ramp by editing DF_Basefile/*.file
exp((seq(from=log(1),to=log(29),length=8)))-1#chlext
}
}
\author{
Joseph Stachelek
}

