% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{grassmap}
\alias{grassmap}
\title{Publication quality maps with GRASS}
\usage{
grassmap(rnge = c(201502), params = c("sal"), fdir = getOption("fdir"),
  basin = "full")
}
\arguments{
\item{rnge}{numeric string of no more than two dates in yyyymm format}

\item{params}{character. string of parameter fields to plot}

\item{fdir}{character file path to local data directory}
}
\value{
output plots to the QGIS_plotting folder
}
\description{
not finished yet
}
\details{
probably need to implement this as a seperate package
}
\examples{
\dontrun{qmap(rnge=c(201502),params=c("sal"))}
}
\author{
Joseph Stachelek
}

