% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{grassmap}
\alias{grassmap}
\title{Publication quality maps with GRASS}
\usage{
grassmap(fpath = NULL, rnge = NULL, params, mapextent = NA,
  numrow = NULL, numcol = NULL, fdir = getOption("fdir"),
  basin = "full", label_string = NULL, labelling = TRUE,
  print_track = FALSE, cleanup = TRUE, rotated = TRUE)
}
\arguments{
\item{fpath}{file.path to geotiff file}

\item{rnge}{numeric string of 1, 2, or more dates dates in yyyymm format. A length 1 rnge will produce a single plot, a length 2 rnge will produce a series of plots bookended by the two dates, a rnge object with more than 2 dates will produce a series of plots exactly corresponding to the dates provided.}

\item{params}{character vector of parameter fields to plot legends and color ramps are defined for sal, chlext, and diffsal}

\item{mapextent}{numeric vector of length 4}

\item{numrow}{numeric number of rows in a multipanel}

\item{numcol}{numeric number of columns in a multipanel}

\item{fdir}{character file path to local data directory}

\item{basin}{character basin name from fathom_basins_proj.shp}

\item{label_string}{character label}

\item{labelling}{logical lablel output with yearmon?}

\item{print_track}{logical print dataflow track?}

\item{cleanup}{logical remove intermediate rasters and shapefiles?}

\item{rotated}{logical rotate canvas to fit Florida Bay more squarely? This requires the i.rotate extension to be installed and addons configured (not working).}
}
\value{
output plots to the QGIS_plotting folder
}
\description{
not finished yet
}
\details{
Probably need to implement this as a seperate package to improve portability. Set param to "diffsal" to plot outpot of avmap function. Will output an imagemagick plot to the working directory and a pdf plot to the file.path(getOption("fdir"), "QGIS_Plotting") folder. The optional fpath argument only supports pointing to a single geotiff.
}
\examples{
\dontrun{
#list supported parameters
grassmap(rnge = c(201512), params = c("sal"))
grassmap(rnge = c(201512), params = c("diffsal"))
grassmap(rnge = c(201407), params = c("chlext"))

#multiple survey dates
grassmap(rnge = c(201509, 201512), params = c("sal"))

#zoom to Manatee + Barnes
grassmap(rnge = c(201512), params = c("sal"), basin = "Manatee Bay")
grassmap(rnge = c(201507), params = c("sal"), mapextent = c(494952.6, 564517.2, 2758908, 2799640))
grassmap(rnge = c(201509), params = c("chlext"), mapextent = c(494952.6,
564517.2, 2758908, 2799640))
grassmap(rnge = c(201512), params = "diffsal", mapextent = c(494952.6, 564517.2, 2758908, 2799640))

#print survey track
grassmap(201513, "chlext", mapextent = c(557217, 567415, 2786102, 2797996), print_track = TRUE)

#specify raster file directly
grassmap(fpath = file.path(getOption("fdir"), "DF_Surfaces", "200904", "sal.tif"), params = "sal")

#specify label string
grassmap(fpath = file.path("/home/jose/Documents/Science",
"sfwmd_desktop/Presentations/2016-02-04_C-111_interagency-monitoring",
 "pre.proj_mean.tif"), params = "sal", label_string = "Pre C-111")

#create a new color ramp by editing DF_Basefile/*.file and update figure makefile
logramp(n = 9, maxrange = 20) #chlext
scales::show_col(viridis::viridis_pal()(9))
grassmap(rnge = c(201509, 201512), params = "sal", numrow = 2, numcol = 1)
}
}
\author{
Joseph Stachelek
}

