% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logramp}
\alias{logramp}
\title{Create a log scaled color ramp for the grassmap function}
\usage{
logramp(n, minrange = NULL, maxrange)
}
\arguments{
\item{n}{integer number of breaks}

\item{minrange}{integer minimum range
n <- 9
maxrange <- 20
logramp(n, maxrange)
scales::show_col(viridis::viridis_pal()(9))}

\item{maxrange}{integer maximum range}
}
\description{
Create a log scaled color ramp for the grassmap function
}
