% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamclean}
\alias{streamclean}
\title{Cleaning raw streaming Dataflow output}
\usage{
streamclean(yearmon, mmin, c6mmin = NA, c6pres = TRUE, eummin = NA,
  eupres = FALSE, exommin = NA, exopres = FALSE, tofile = FALSE,
  sep = ",", fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric designation of survey date formatted as yyyymm}

\item{mmin}{minimum measurement frequency (# measurements/min)}

\item{c6mmin}{integer optional minimum c6 measurement frequency if different than mmin}

\item{c6pres}{logical was C6 data collected?}

\item{eummin}{integer optional minimum eureka measurement frequency if different than mmin}

\item{eupres}{logical was eureka data collected?}

\item{exommin}{integer optional minimum exo measurement frequency if different than mmin}

\item{exopres}{logical was exo data collected?}

\item{tofile}{logical save cleaned output to DF_FullDataSets?}

\item{sep}{character optional predesignation of item seperation character in raw data files}

\item{fdir}{character file path to local data directory}
}
\description{
Cleaning raw streaming Dataflow, C6, Eureka-Manta, and YSI-Exo output
}
\details{
Dataflow cleaning drops all minutes that have less measurements than "mmin". C6 data is interpolated to match Dataflow.  Automatically compares salinity against conducitivty/temperature recalculated salinity and replaces if slope of fit is not close to 1. Bad DO columns must sometimes be removed manually. TODO - Add check the make sure that the year of the data (not just the filename) matches the year of yearmon
}
\examples{
\dontrun{
dt <- streamclean(yearmon = 201505, mmin = 7, c6mmin = 10, tofile = FALSE, c6pres = TRUE)
dt <- streamclean(yearmon = 201513, mmin = 7, c6pres = TRUE, c6mmin = 12,
tofile = FALSE, exommin = 60, exopres = TRUE, eupres = TRUE, eummin = 12)
}
}

