% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamclean}
\alias{streamclean}
\title{Cleaning raw streaming Dataflow output}
\usage{
streamclean(yearmon, gps, dfmmin = NA, c6mmin = NA, eummin = NA,
  exommin = NA, tofile = FALSE, sep = ",", fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric designation of survey date formatted as yyyymm}

\item{gps}{character dataset to use for GPS alignment of other data streams. Choice of "df", "eu", or "exo".}

\item{dfmmin}{integer optional minimum df measurement frequency (# measurements/min)}

\item{c6mmin}{integer optional minimum c6 measurement frequency}

\item{eummin}{integer optional minimum eureka (manta) measurement frequency}

\item{exommin}{integer optional minimum exo measurement frequency}

\item{tofile}{logical save cleaned output to DF_FullDataSets?}

\item{sep}{character optional predesignation of item seperation character in raw data files}

\item{fdir}{character file path to local data directory}
}
\description{
Cleaning raw streaming Dataflow, C6, Eureka-Manta, and YSI-Exo output
}
\details{
Dataflow cleaning drops all minutes that have less measurements than "mmin". C6 data is interpolated to match Dataflow.  Automatically compares salinity against conducitivty/temperature recalculated salinity and replaces if slope of fit is not close to 1. Bad DO columns must sometimes be removed manually. TODO - Add check the make sure that the year of the data (not just the filename) matches the year of yearmon
}
\examples{
\dontrun{
#old
dt <- streamclean(yearmon = 201505, dfmmin = 7, c6mmin = 10,
tofile = FALSE)
dt <- streamclean(yearmon = 201513, dfmmin = 7, c6mmin = 12,
tofile = FALSE, exommin = 60, eummin = 12)

#working
dt <- streamclean(yearmon = 201512, gps = "df", c6mmin = 6, dfmmin = 7)
dt <- streamclean(yearmon = 201601, gps = "eu", eummin = 12)
dt <- streamclean(yearmon = 201603, gps = "exo", exommin = 40, c6mmin = 12)
}
}

