% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamget}
\alias{streamget}
\title{Retrieve previously cleaned full streaming datasets}
\usage{
streamget(yearmon, fdir)
}
\arguments{
\item{yearmon}{numeric date in yyyymm format}

\item{fdir}{character file path to local data directory}
}
\description{
Retrieve previously cleaned full streaming datasets
}
\examples{
\dontrun{
yearmon<-201308
dt<-streamget(yearmon)}
}

