% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamget}
\alias{streamget}
\title{Retrieve previously cleaned full streaming datasets}
\usage{
streamget(yearmon, qa = TRUE, fdir = getOption("fdir"))
}
\arguments{
\item{yearmon}{numeric date in yyyymm format}

\item{qa}{logical strip flagged data?}

\item{fdir}{character file path to local data directory}
}
\description{
Retrieve previously cleaned full streaming datasets
}
\examples{
\dontrun{
yearmon <- 201212
dt <- streamget(yearmon)
}
}
