% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streaminginterp.R
\name{streaminterp}
\alias{streaminterp}
\title{interpolation of streaming data}
\usage{
streaminterp(dt, paramlist, yearmon, trim_rstack = TRUE,
  costrasname = "barrier60large2e2.tif", tname = NA, vname = NA,
  missprop = 0.16, fdir = getOption("fdir"))
}
\arguments{
\item{dt}{input data frame}

\item{paramlist}{list of parameters (dt column names) to interpolate}

\item{yearmon}{a file path used to extract basename}

\item{trim_rstack}{logical trim the raster stack by the convex hull of training points?}

\item{tname}{file.path location to save training dataset}

\item{vname}{file.path location to save validation dataset}

\item{missprop}{numeric proportion of missing data allowed. Variables with a greater proportion of missing data will be dropped.}

\item{fdir}{character file path to local data directory}
}
\examples{
\dontrun{
dt <- streamget(yearmon = 201513, qa = TRUE)
streaminterp(dt, paramlist = c("sal"), yearmon = 201513)
}
}

