% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/streamingclean.R
\name{streamqa}
\alias{streamqa}
\title{Retrieve previously cleaned full streaming datasets}
\usage{
streamqa(yearmon, setthresh = TRUE, trimends = TRUE, paired = FALSE)
}
\arguments{
\item{yearmon}{numeric date in yyyymm format}

\item{setthresh}{logical set parameter thresholds}

\item{trimends}{logical look to trim ends of data stream?}

\item{paired}{logical examine relationships between paried parameters?}
}
\value{
a matrix of the same size/shape of the fulldataset, with entries specifying where to set to NA, saved to DF_FullDataSets/Raw/IntrumentOutput
}
\description{
Retrieve previously cleaned full streaming datasets
}
\details{
loop through parameters giving the opportunity to trim measurement ends, set entire variables to NA, remove variables above/below a threshold
}
\examples{
\dontrun{
dt<-streamqa(yearmon=201410)}
}

