% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{surfplot}
\alias{surfplot}
\title{Plotting Interpolated Surfaces}
\usage{
surfplot(rnge = c(201402, 201404), params = c("c6chl", "sal"),
  fdir = getOption("fdir"))
}
\arguments{
\item{rnge}{numeric string of no more than two dates in yyyymm format}

\item{params}{character. string of parameter fields to plot}

\item{fdir}{character file path to local data directory}
}
\value{
output plots to plot window
}
\description{
accounts for corner case parameter spellings, variable specific contour breaks are (need to be) defined
}
\examples{
\dontrun{
surfplot(rnge=c(200707),params=c("cdom"))
}
}
\author{
Joseph Stachelek
}

