% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surfaceplot.R
\name{surfplot}
\alias{surfplot}
\title{Plotting Interpolated Surfaces}
\usage{
surfplot(rnge = c(201402, 201404), params = c("c6chl", "sal"),
  fdir = getOption("fdir"), yext = c(2772256, 2798000), xext = c(518000.2,
  566000))
}
\arguments{
\item{rnge}{numeric string of no more than two dates in yyyymm format}

\item{params}{character. string of parameter fields to plot}

\item{fdir}{character file path to local data directory}

\item{yext}{numeric length 2 y extent}

\item{xext}{numeric length 2 x extent}
}
\value{
output plots to plot window
}
\description{
accounts for corner case parameter spellings, variable specific contour breaks are (need to be) defined
}
\examples{
\dontrun{
surfplot(rnge=c(200707), params = c("cdom"))
surfplot(201513, c("ph", "c6turbidity", "c6chl", "c6cdom"),
yext = c(2786102, 2797996), xext = c(557217, 567415))
}
}
\author{
Joseph Stachelek
}

